==> 0_subway_day.ipynb <==
import pandas as pd
import glob
import json
import requests

# 지하철 역별 위치 정보 저장 (위도, 경도, 행정구역 명)

# `sub_yx` 지하철 위경도 좌표가 적힌 df 생성
sub_yx = pd.read_csv('data/subway_day/지하철역_좌표.csv', encoding='cp949')

# `sub_yx` df에 좌표데이터와 매칭되는 행정구역 명 추가
api_key = '2c56f9597f650d4a2dec845b2e0fe615'

# 위도 경도로 주소를 찾아서 리스트에 넣어주는 함수
def lat_lon_to_addr(lon,lat):
    url = 'https://dapi.kakao.com/v2/local/geo/coord2regioncode.json?x={longitude}&y={latitude}'.format(longitude=lon,latitude=lat)
    headers = {"Authorization": "KakaoAK " + api_key}
    result = json.loads(str(requests.get(url, headers=headers).text))
    match_first = result['documents'][0]['address_name']
    return addr.append(match_first)

addr = []

for i in range(0,len(sub_yx['X좌표'])):
    try: lat_lon_to_addr(sub_yx['X좌표'][i],sub_yx['Y좌표'][i])
    except :
        addr.append(None)
        
# #리스트를 데이터 프레임으로 저장
addr = pd.DataFrame(addr)

#컬럼 정리
addr['X좌표'] = sub_yx['X좌표']
addr.rename(columns={0:'주소'},inplace=True)
addr=addr[['X좌표', '주소']]

#병합하고 분리
sub_yx = pd.merge(sub_yx,addr, on = 'X좌표', how='left')

sub_yx = sub_yx[sub_yx['주소'].str.split(' ').str.get(0) == '서울특별시']

sub_yx['구'] = sub_yx['주소'].str.split(' ').str.get(-2)
sub_yx['동'] = sub_yx['주소'].str.split(' ').str.get(-1)


# 저장해서 로드 시간을 단축
sub_yx.to_csv('data/subway_day/지하철역_위치데이터.csv', index=False)

### (지하철 위치데이터 `sub_yx` 데이터프레임 정보)

type(sub_yx)

sub_yx.info()

sub_yx.dtypes

sub_yx.size

sub_yx.ndim

sub_yx.index

sub_yx.values

sub_yx.columns

sub_yx.head(2)

sub_yx.tail(2)

sub_yx['역명'].nunique()

# 2017년_지하철데이터 생성

##  1. 2017년도 월별 지하철 지하철탑승정보 csv 파일 병합

# csv 파일을 불러 병합
csvfiles = glob.glob('data/subway_day/2017/*.csv')
# `sub_off2017` 2017년도 지하철 하차인원 csv 파일의 합 df 생성
sub_off2017 = pd.DataFrame()

for csvfile in csvfiles:
    tmp = pd.read_csv(csvfile, encoding='cp949', low_memory = False)
    sub_off2017 = pd.concat([sub_off2017, tmp])
    
sub_off2017.reset_index(drop=True, inplace=True)

### (`sub_off2017` 데이터프레임 정보)

type(sub_off2017)

sub_off2017.info()

sub_off2017.dtypes

sub_off2017.size

sub_off2017.ndim

sub_off2017.index

sub_off2017.values

sub_off2017.columns

sub_off2017.head(2)

sub_off2017.tail(2)

# 총 지하철역 개수 파악
sub_off2017['역명'].nunique()

## 2. 데이터 정제 및 저장

# 불필요한 칼럼 제거
sub_off2017.drop(['노선명', '승차총승객수', '등록일자'], axis=1, inplace=True)

# `역명` 칼럼을 기준으로 sub_off2017에 sub_yx에서 역명에 일치하는 위도, 경도값, 행정구역 명 병합
sub_2017 = sub_off2017.merge(sub_yx, on='역명', how='left')

# 주소값이 매칭되지 않아서 서울시 이외의 `역명` 컬럼에 null 값이 들어간 행 제거
sub_2017.dropna(axis=0, inplace = True)

sub_2017.head(2)

sub_2017.info()

sub_2017['역명'].nunique()

# `int64` 타입이였던 `사용일자` 컬럼의 데이터들을 `string` 타입으로 변환 한 후에 
# 인덱싱을 통해 `년`, `월`, `일` 형태에 맞게 컬럼으로 추가
sub_2017['년'] = sub_2017['사용일자'].astype('string').str[:4]
sub_2017['월'] = sub_2017['사용일자'].astype('string').str[4:6]
sub_2017['일'] = sub_2017['사용일자'].astype('string').str[6:]

# 컬럼의 순서를 다른 데이터와 합치기 좋게 배치
sub_2017 = sub_2017[['사용일자', '년', '월', '일', '역명', '하차총승객수', 'X좌표', 'Y좌표', '주소', '구', '동']]

# `역명` 컬럼에 혹시나 좌표값이 들어가지 않은 열이 있는지 `isnull`을 통해서 검토
sub_2017['역명'].isnull().sum()

# 만약 `역명` 컬럼에 `null`이 있는지 `isnull`로 개수를 알아봤다면
# `null` 값이 들어간 해당 열이 어딘지 확인
# sub_2017[sub_2017['역명'].isnull()]

# 정제가 끝난 후 파일을 마지막으로 df 검토
sub_2017.head()

# 정제한 데이터프레임을 csv 파일 형식으로 저장
sub_2017.to_csv('data/subway_total/2017년_지하철데이터.csv',index=False)

# 2018년 지하철데이터 생성

- ( 2018 ~ 2021 년도 ) 데이터 정제 방식은 위의 2017 년도의 과정과 모두 동일

- 2018 년도 `서울역` 승하차 인원 데이터만 원본 데이터에도 기록이 제대로 안된 부분들이 있음

csvfiles = glob.glob('data/subway_day/2018/*.csv')
sub_off2018 = pd.DataFrame()

for csvfile in csvfiles:
    tmp = pd.read_csv(csvfile, encoding='cp949', low_memory = False)
    sub_off2018 = pd.concat([sub_off2018, tmp])
    
sub_off2018.reset_index(drop=True, inplace=True)

sub_off2018.drop(['노선명', '승차총승객수', '등록일자'], axis=1, inplace=True)

sub_2018 = sub_off2018.merge(sub_yx, on='역명')

sub_2018.dropna(axis=0, inplace = True)

sub_2018['년'] = sub_2018['사용일자'].astype('string').str[:4]
sub_2018['월'] = sub_2018['사용일자'].astype('string').str[4:6]
sub_2018['일'] = sub_2018['사용일자'].astype('string').str[6:]

sub_2018 = sub_2018[['사용일자', '년', '월', '일', '역명', '하차총승객수', 'X좌표', 'Y좌표', '주소', '구', '동']]

sub_2018.to_csv('data/subway_total/2018년_지하철데이터.csv',index=False)

# 2019년 지하철데이터 생성

csvfiles = glob.glob('data/subway_day/2019/*.csv')
sub_off2019 = pd.DataFrame()

for csvfile in csvfiles:
    tmp = pd.read_csv(csvfile, encoding='cp949', low_memory = False)
    sub_off2019 = pd.concat([sub_off2019, tmp])
    
sub_off2019.reset_index(drop=True, inplace=True)

sub_off2019.drop(['노선명', '승차총승객수', '등록일자'], axis=1, inplace=True)

sub_2019 = sub_off2019.merge(sub_yx, on='역명')

sub_2019.dropna(axis=0, inplace = True)

sub_2019['년'] = sub_2019['사용일자'].astype('string').str[:4]
sub_2019['월'] = sub_2019['사용일자'].astype('string').str[4:6]
sub_2019['일'] = sub_2019['사용일자'].astype('string').str[6:]

sub_2019 = sub_2019[['사용일자', '년', '월', '일', '역명', '하차총승객수', 'X좌표', 'Y좌표', '주소', '구', '동']]

sub_2019.to_csv('data/subway_total/2019년_지하철데이터.csv',index=False)

# 2020년 지하철데이터 생성

csvfiles = glob.glob('data/subway_day/2020/*.csv')
sub_off2020 = pd.DataFrame()

for csvfile in csvfiles:
    tmp = pd.read_csv(csvfile, encoding='cp949', low_memory = False)
    sub_off2020 = pd.concat([sub_off2020, tmp])
    
sub_off2020.reset_index(drop=True, inplace=True)

sub_off2020.drop(['노선명', '승차총승객수', '등록일자'], axis=1, inplace=True)

sub_2020 = sub_off2020.merge(sub_yx, on='역명')

sub_2020.dropna(axis=0, inplace = True)

sub_2020['년'] = sub_2020['사용일자'].astype('string').str[:4]
sub_2020['월'] = sub_2020['사용일자'].astype('string').str[4:6]
sub_2020['일'] = sub_2020['사용일자'].astype('string').str[6:]

sub_2020 = sub_2020[['사용일자', '년', '월', '일', '역명', '하차총승객수', 'X좌표', 'Y좌표', '주소', '구', '동']]

sub_2020.to_csv('data/subway_total/2020년_지하철데이터.csv',index=False)

# 2021년 지하철데이터 생성

csvfiles = glob.glob('data/subway_day/2021/*.csv')
sub_off2021 = pd.DataFrame()

for csvfile in csvfiles:
    tmp = pd.read_csv(csvfile, encoding='cp949', low_memory = False)
    sub_off2021 = pd.concat([sub_off2021, tmp])
    
sub_off2021.reset_index(drop=True, inplace=True)

sub_off2021.drop(['노선명', '승차총승객수', '등록일자'], axis=1, inplace=True)

sub_2021 = sub_off2021.merge(sub_yx, on='역명')

sub_2021.dropna(axis=0, inplace = True)

sub_2021['년'] = sub_2021['사용일자'].astype('string').str[:4]
sub_2021['월'] = sub_2021['사용일자'].astype('string').str[4:6]
sub_2021['일'] = sub_2021['사용일자'].astype('string').str[6:]

sub_2021 = sub_2021[['사용일자', '년', '월', '일', '역명', '하차총승객수', 'X좌표', 'Y좌표', '주소', '구', '동']]

sub_2021.to_csv('data/subway_total/2021년_지하철데이터.csv',index=False)
