==> 1_combine_bus_sub.ipynb <==
import pandas as pd

data_bus = pd.read_csv('data/bus_total/2017버스.csv')
data_bus.head(3)

data_bus.rename(columns={'버스정류장ARS번호':'식별'}, inplace=True)

data_bus['구분'] = '버스'

data_sub = pd.read_csv('data/subway_total/2017년_지하철데이터.csv')
data_sub.head(3)

data_sub.rename(columns={'역명':'식별'}, inplace=True)

data_sub['구분'] = '지하철'

data = pd.concat([data_bus, data_sub])
data.head(3)

data.tail(3)

data.info()

data.to_csv('data/total/total_2017.csv', index=False)

# 2018 합치기

data_bus = pd.read_csv('data/bus_total/2018버스.csv')

data_bus.rename(columns={'버스정류장ARS번호':'식별'}, inplace=True)

data_bus['구분'] = '버스'

data_sub = pd.read_csv('data/subway_total/2018년_지하철데이터.csv')

data_sub.rename(columns={'역명':'식별'}, inplace=True)

data_sub['구분'] = '지하철'

data = pd.concat([data_bus, data_sub])

data.head(3)

data.tail(3)

data.to_csv('data/total/total_2018.csv', index=False)

# 2019 합치기

data_bus = pd.read_csv('data/bus_total/2019버스.csv')

data_bus.rename(columns={'버스정류장ARS번호':'식별'}, inplace=True)

data_bus['구분'] = '버스'

data_sub = pd.read_csv('data/subway_total/2019년_지하철데이터.csv')

data_sub.rename(columns={'역명':'식별'}, inplace=True)

data_sub['구분'] = '지하철'

data = pd.concat([data_bus, data_sub])

data.head(3)

data.tail(3)

data.to_csv('data/total/total_2019.csv', index=False)

# 2020 합치기

data_bus = pd.read_csv('data/bus_total/2020버스.csv')

data_bus.rename(columns={'버스정류장ARS번호':'식별'}, inplace=True)

data_bus['구분'] = '버스'

data_sub = pd.read_csv('data/subway_total/2020년_지하철데이터.csv')

data_sub.rename(columns={'역명':'식별'}, inplace=True)

data_sub['구분'] = '지하철'

data = pd.concat([data_bus, data_sub])

data.head(3)

data.tail(3)

data.to_csv('data/total/total_2020.csv', index=False)

# 2021 합치기

data_bus = pd.read_csv('data/bus_total/2021버스.csv')

data_bus.rename(columns={'버스정류장ARS번호':'식별'}, inplace=True)

data_bus['구분'] = '버스'

data_sub = pd.read_csv('data/subway_total/2021년_지하철데이터.csv')

data_sub.rename(columns={'역명':'식별'}, inplace=True)

data_sub['구분'] = '지하철'

data = pd.concat([data_bus, data_sub])

data.head(3)

data.tail(3)

data.to_csv('data/total/total_2021.csv', index=False)
