==> 2_calender.ipynb <==
import pyecharts
from pyecharts.charts import Calendar
from pyecharts import options as opts
import pandas as pd

# 2017

total_2017 = pd.read_csv('data/total/total_2017.csv')

total_2017['사용일자'] = total_2017['사용일자'].astype('str')
total_2017['사용일자'] = pd.to_datetime(total_2017['사용일자'])
day_by_year = total_2017.groupby('사용일자').sum()[['하차총승객수']].round(0)
day_by_year = day_by_year.reset_index()
data = day_by_year[['사용일자','하차총승객수']].values.tolist()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
cal = (Calendar()
 .add('',data,calendar_opts=opts.CalendarOpts(range_='2017'))
 .set_global_opts(
     title_opts=opts.TitleOpts(title='2017년도 지하철, 버스 유동인구', subtitle='일별 데이터'),
     legend_opts=opts.LegendOpts(is_show=False),
     visualmap_opts=opts.VisualMapOpts(
         max_ = max_day,
         min_ = min_day,
         orient='horizontal',
         is_piecewise=False,
         pos_top='230px',
         pos_left='100px',
     )))

cal.render('data/total/calender/2017render.html') #생성된 파일 확인

# 2018

total_2018 = pd.read_csv('data/total/total_2018.csv')

total_2018['사용일자'] = total_2018['사용일자'].astype('str')
total_2018['사용일자'] = pd.to_datetime(total_2018['사용일자'])
day_by_year = total_2018.groupby('사용일자').sum()[['하차총승객수']].round(0)
day_by_year = day_by_year.reset_index()
data = day_by_year[['사용일자','하차총승객수']].values.tolist()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
cal = (Calendar()
 .add('',data,calendar_opts=opts.CalendarOpts(range_='2018'))
 .set_global_opts(
     title_opts=opts.TitleOpts(title='2018년도 지하철, 버스 유동인구', subtitle='일별 데이터'),
     legend_opts=opts.LegendOpts(is_show=False),
     visualmap_opts=opts.VisualMapOpts(
         max_ = max_day,
         min_ = min_day,
         orient='horizontal',
         is_piecewise=False,
         pos_top='230px',
         pos_left='100px',
     )))

cal.render('data/total/calender/2018render.html') #생성된 파일 확인

# 2019

total_2019 = pd.read_csv('data/total/total_2019.csv')

total_2019['사용일자'] = total_2019['사용일자'].astype('str')
total_2019['사용일자'] = pd.to_datetime(total_2019['사용일자'])
day_by_year = total_2019.groupby('사용일자').sum()[['하차총승객수']].round(0)
day_by_year = day_by_year.reset_index()
data = day_by_year[['사용일자','하차총승객수']].values.tolist()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
cal = (Calendar()
 .add('',data,calendar_opts=opts.CalendarOpts(range_='2019'))
 .set_global_opts(
     title_opts=opts.TitleOpts(title='2019년도 지하철, 버스 유동인구', subtitle='일별 데이터'),
     legend_opts=opts.LegendOpts(is_show=False),
     visualmap_opts=opts.VisualMapOpts(
         max_ = max_day,
         min_ = min_day,
         orient='horizontal',
         is_piecewise=False,
         pos_top='230px',
         pos_left='100px',
     )))

cal.render('data/total/calender/2019render.html') #생성된 파일 확인

# 2020

total_2020 = pd.read_csv('data/total/total_2020.csv')

total_2020['사용일자'] = total_2020['사용일자'].astype('str')
total_2020['사용일자'] = pd.to_datetime(total_2020['사용일자'])
day_by_year = total_2020.groupby('사용일자').sum()[['하차총승객수']].round(0)
day_by_year = day_by_year.reset_index()
data = day_by_year[['사용일자','하차총승객수']].values.tolist()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
cal = (Calendar()
 .add('',data,calendar_opts=opts.CalendarOpts(range_='2020'))
 .set_global_opts(
     title_opts=opts.TitleOpts(title='2020 지하철, 버스 유동인구', subtitle='일별 데이터'),
     legend_opts=opts.LegendOpts(is_show=False),
     visualmap_opts=opts.VisualMapOpts(
         max_ = max_day,
         min_ = min_day,
         orient='horizontal',
         is_piecewise=False,
         pos_top='230px',
         pos_left='100px',
     )))

cal.render('data/total/calender/2020render.html') #생성된 파일 확인

# 2021

total_2021 = pd.read_csv('data/total/total_2021.csv')

total_2021['사용일자'] = total_2021['사용일자'].astype('str')
total_2021['사용일자'] = pd.to_datetime(total_2021['사용일자'])
day_by_year = total_2021.groupby('사용일자').sum()[['하차총승객수']].round(0)
day_by_year = day_by_year.reset_index()
data = day_by_year[['사용일자','하차총승객수']].values.tolist()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
max_day = day_by_year['하차총승객수'].max()
min_day = day_by_year['하차총승객수'].min()
cal = (Calendar()
 .add('',data,calendar_opts=opts.CalendarOpts(range_='2021'))
 .set_global_opts(
     title_opts=opts.TitleOpts(title='2021년도 지하철, 버스 유동인구', subtitle='일별 데이터'),
     legend_opts=opts.LegendOpts(is_show=False),
     visualmap_opts=opts.VisualMapOpts(
         max_ = max_day,
         min_ = min_day,
         orient='horizontal',
         is_piecewise=False,
         pos_top='230px',
         pos_left='100px',
     )))

cal.render('data/total/calender/2021render.html') #생성된 파일 확인
